Refactoring combat listener in preparation for re-using damage calculators
