+++ src/main/java/info/faceland/strife/StrifePlugin.java
@@ -333,6 +333,7 @@ public class StrifePlugin extends FacePlugin {
     Bukkit.getPluginManager().registerEvents(new HeadDropListener(this), this);
     Bukkit.getPluginManager().registerEvents(new DataListener(this), this);
     Bukkit.getPluginManager().registerEvents(new AttributeUpdateListener(this), this);
+    Bukkit.getPluginManager().registerEvents(new DogeListener(entityStatCache), this);
     if (Bukkit.getPluginManager().getPlugin("Bullion") != null) {
       Bukkit.getPluginManager().registerEvents(new BullionListener(this), this);
     }
